+++ packages/toisu/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## v6.0.0
+
+- The lowest supported version of Node is now 18.
+- The default error handler now looks for a `statusCode` property on the error
+  it receives, and uses it to set the status code of the response. Many third
+  party modules (and first party wrappers like the Toisu! static module) produce
+  such errors.
+- Not-found handling can now be customized by setting a `handleNotFound`
+  property on Toisu! objects.
+
 ## v5.0.0
 
 - The lowest supported version of Node is now 16.